Add TAR_STREAM and TAR_GZ_STREAM modes to extract from .tar and tar.gz streams on-the-fly.
The new modes are faster and can use less temporary disk/memory than TAR and TAR_GZ, but random access to the files is not supported.

In order to use these modes, one has to make them explicit, using the Resource's extract_method kwarg.

The default mode for extracting .tar and .tar.gz has not changed. The only difference is that getmembers() has been replaced by an iterator.

The results of some benchmarks that I ran on my workstation, iterating through a 6GB tar.gz (on a SSD disk):
- Old: time=0:33.72, resmem=51420KB, maxdisk=~720KB
- New: time=0:33.47, resmem=53224KB, maxdisk=~750KB
- New+stream: time=0:16.30, resmem=53280KB, maxdisk=~250KB

PiperOrigin-RevId: 240627378